Material IO: Rename and split importer base into separate file
