+++ package/com.unity.formats.usd/Runtime/Scripts/IO/Materials/StandardShaderImporter.cs
@@ -13,128 +13,13 @@
 // limitations under the License.
 
 using System;
-using System.Collections.Generic;
 using UnityEngine;
 
 namespace USD.NET.Unity {
-  public abstract class MaterialAdapter {
-    public Material Material { get; private set; }
-    public bool IsMetallicWorkflow { get; private set; }
 
-    public Color? Diffuse;
-    public Texture2D DiffuseMap;
+  public class StandardShaderImporter : MaterialImporterBase {
 
-    public Color? Specular;
-    public Texture2D SpecularMap;
-
-    public Vector3? Normal;
-    public Texture2D NormalMap;
-
-    public float? Displacement;
-    public Texture2D DisplacementMap;
-
-    public float? Occlusion;
-    public Texture2D OcclusionMap;
-
-    public Color? Emission;
-    public Texture2D EmissionMap;
-
-    public float? Roughness;
-    public Texture2D RoughnessMap;
-
-    public float? Metallic;
-    public Texture2D MetallicMap;
-
-    public MaterialAdapter(Material material) {
-      Material = material;
-    }
-
-    abstract public void ImportFromUsd();
-    abstract public void ExportToUsd();
-
-    protected void ImportColorOrMap(Scene scene,
-                                    Connectable<Vector3> usdParam,
-                                    bool isNormalMap,
-                                    SceneImportOptions options,
-                                    ref Texture2D map,
-                                    ref Color? value,
-                                    out string uvPrimvar) {
-      uvPrimvar = null;
-      if (usdParam.IsConnected()) {
-        map = MaterialImporter.ImportConnectedTexture(scene, usdParam, isNormalMap, options, out uvPrimvar);
-      } else {
-        var rgb = usdParam.defaultValue;
-        value = new Color(rgb.x, rgb.y, rgb.z).gamma;
-      }
-    }
-
-    protected void ImportValueOrMap<T>(Scene scene,
-                                    Connectable<T> usdParam,
-                                    bool isNormalMap,
-                                    SceneImportOptions options,
-                                    ref Texture2D map,
-                                    ref T? value,
-                                    out string uvPrimvar) where T: struct {
-      uvPrimvar = null;
-      if (usdParam.IsConnected()) {
-        map = MaterialImporter.ImportConnectedTexture(scene, usdParam, isNormalMap, options, out uvPrimvar);
-      } else {
-        value = usdParam.defaultValue;
-      }
-    }
-
-    private void MergePrimvars(string newPrimvar, List<string> primvars) {
-      if (string.IsNullOrEmpty(newPrimvar)) {
-        return;
-      }
-      string localName = newPrimvar;
-      if (!string.IsNullOrEmpty(primvars.Find(str => str == localName))) {
-        return;
-      }
-      primvars.Add(localName);
-    }
-
-    public virtual void ImportParametersFromUsd(Scene scene,
-                                                string materialPath,
-                                                MaterialSample materialSample,
-                                                PreviewSurfaceSample previewSurf,
-                                                SceneImportOptions options) {
-      var primvars = new List<string>();
-      string uvPrimvar = null;
-      ImportColorOrMap(scene, previewSurf.diffuseColor, false, options, ref DiffuseMap, ref Diffuse, out uvPrimvar);
-      MergePrimvars(uvPrimvar, primvars);
-
-      ImportColorOrMap(scene, previewSurf.emissiveColor, false, options, ref EmissionMap, ref Emission, out uvPrimvar);
-      MergePrimvars(uvPrimvar, primvars);
-
-      ImportValueOrMap(scene, previewSurf.normal, true, options, ref NormalMap, ref Normal, out uvPrimvar);
-      MergePrimvars(uvPrimvar, primvars);
-
-      ImportValueOrMap(scene, previewSurf.displacement, false, options, ref DisplacementMap, ref Displacement, out uvPrimvar);
-      MergePrimvars(uvPrimvar, primvars);
-
-      ImportValueOrMap(scene, previewSurf.occlusion, false, options, ref OcclusionMap, ref Occlusion, out uvPrimvar);
-      MergePrimvars(uvPrimvar, primvars);
-
-      ImportValueOrMap(scene, previewSurf.roughness, false, options, ref RoughnessMap, ref Roughness, out uvPrimvar);
-      MergePrimvars(uvPrimvar, primvars);
-
-      if (previewSurf.useSpecularWorkflow.defaultValue == 1) {
-        ImportColorOrMap(scene, previewSurf.specularColor, false, options, ref SpecularMap, ref Specular, out uvPrimvar);
-        MergePrimvars(uvPrimvar, primvars);
-      } else {
-        ImportValueOrMap(scene, previewSurf.metallic, false, options, ref MetallicMap, ref Metallic, out uvPrimvar);
-        MergePrimvars(uvPrimvar, primvars);
-      }
-
-      options.materialMap.SetPrimvars(materialPath, primvars);
-    }
-
-  }
-
-  public class StandardShaderAdapter : MaterialAdapter {
-
-    public StandardShaderAdapter(Material material) : base(material) {
+    public StandardShaderImporter(Material material) : base(material) {
     }
 
     public override void ExportToUsd() {
